requirements.txt
# Web Framework
streamlit>=1.18.0

# REST API
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Data Science & Machine Learning
pandas>=1.5.0
numpy>=1.23.0
scikit-learn==1.7.2
joblib>=1.2.0

# Visualization
matplotlib>=3.6.0
seaborn>=0.12.0
plotly>=5.11.0

# Notebook Support
jupyter>=1.0.0
ipykernel>=6.19.0

# Code Analysis
ast-comments>=1.1.0

# Utilities
python-dotenv>=0.21.0
